Implemented simple notification for buxton keys
A buxton client connection is left open in the simple library as
long as there is at least one notification registered with the
daemon. Calls to simple register take a key name and a callback
function. These, along with the previously set group and layer,
are used to get a key type and register a notification in the
daemon.

A provided callback deals with the BuxtonResponse and calls the
user's callback with the new data as a void * and the key name
as a char *.

sbuxton_get_fd and sbuxton_handle_response are provided to use in
polling or a mainloop.

Includes documentation for all new public functions and passing
checks for new internal functions.