Implemented get_key_type feature in libbuxton, daemon, and buxtonctl
Get_key_type returns the type of a key's value as a uint32_t, which can
be cast to a BuxtonDataType. This commit adds buxton_get_key_type to
libbuxton, adds support functions for it in protocol, daemon, direct,
and gdbm, and adds this feature to buxtonctl.

A demo program, bxt_hello_get_key_type, a man page, buxton_get_key_type,
and changes to other affected man pages are included.

Passing unit tests are in check_daemon and check_buxton.