Implementing the listing of names
The functions ...list_keys are renamed ...list_names
and are now listing either the names of the groups or
of the keys of a group.

Listing layers was considered but rejected for this
version in the reason of its wider impact on the code.

Currently only the gdm version is implemented.

Change-Id: I75fc5a00d45968e57c344faefd0868e27f181bab