Introduce BuxtonHashmap
BuxtonHashmap is a cache aggressive hashmap designed to replace the existing
hashmap implementation in Buxton, ensuring clean operations, high performance
and simple iteration functionality.

This commit is in relation to issue #124 and unit tests will also follow to
ensure decent code coverage, as well as replacing the existing use of the
systemd hashmap throughout Buxton.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>