cli: option to enforce server connection
The option --server (-s) enforces the program
buxtonctl to use a socket connection to the
server and to not try to use the direct API.

This option is needed for memory layers that
are mandatory handled by server.

Also, the command check allow to just check if
the connection is available directly or through
server link depending on --direct and --server
options.

Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>